smarthole/requirements.txt
pandas
numpy
scipy
scikit-learn
matplotlib
seaborn
joblib
pyyaml
